Memory Usage Information for Parallel R Code

This function is a wrapper around the system command ps that can be used to benchmark (peak) memory usage of parallel R code. By taking snapshots the memory usage of R processes at a regular interval, the function dynamically builds up a profile of their usage of system resources.

syrup(expr, interval = 0.5, peak = FALSE, env = caller_env())

Arguments

expr

An expression.

interval

The interval at which to take snapshots of memory usage. In practice, there's an overhead on top of each of these intervals.

peak

Whether to return rows for only the "peak" memory usage. Interpreted as the id with the maximum rss sum. Defaults to FALSE, but may be helpful to set peak = TRUE for potentially very long-running processes so that the tibble doesn't grow too large.

env

The environment to evaluate expr in.

Value

A tibble with column id and a number of columns from ps::ps() output describing memory usage. Notably, the process ID pid, parent process ID ppid, and resident set size rss (a measure of memory usage).

Details

There's nothing specific about this function that necessitates the provided expression is run in parallel. Said another way, syrup() will work just fine with "normal," sequentially-run R code (as in the examples). That said, there are many better, more fine-grained tools for the job in the case of sequential R code, such as Rprofmem(), the profmem package, the bench package, and packages in the R-prof GitHub organization.

Loosely, the function works by:

Setting up another R process (call it sesh) that queries system information using ps::ps() at a regular interval,

Evaluating the supplied expression,

Reading the queried system information back into the main process from sesh,

Closing sesh, and then

Returning the queried system information.

Note that information on the R process sesh is filtered out from the results automatically.

Examples

# pass any expression to syrup. first, sequentially:
res_syrup <- syrup({res_output <- Sys.sleep(1)})

res_syrup
#> # A tibble: 3 × 6
#>      id   pid  ppid name        rss       vms
#>   <dbl> <int> <int> <chr> <bch:byt> <bch:byt>
#> 1     1  5820  1612 R         244MB     903MB
#> 2     2  5820  1612 R         244MB     903MB
#> 3     3  5820  1612 R         244MB     903MB

# to snapshot memory information more (or less) often, set `interval`
syrup(Sys.sleep(1), interval = .01)
#> # A tibble: 32 × 6
#>       id   pid  ppid name        rss       vms
#>    <dbl> <int> <int> <chr> <bch:byt> <bch:byt>
#>  1     1  5820  1612 R         248MB     907MB
#>  2     2  5820  1612 R         248MB     907MB
#>  3     3  5820  1612 R         248MB     907MB
#>  4     4  5820  1612 R         248MB     907MB
#>  5     5  5820  1612 R         248MB     907MB
#>  6     6  5820  1612 R         248MB     907MB
#>  7     7  5820  1612 R         248MB     907MB
#>  8     8  5820  1612 R         248MB     907MB
#>  9     9  5820  1612 R         248MB     907MB
#> 10    10  5820  1612 R         248MB     907MB
#> # ℹ 22 more rows

# use `peak = TRUE` to return only the snapshot with
# the highest memory usage (as `sum(rss)`)
syrup(Sys.sleep(1), interval = .01, peak = TRUE)
#> # A tibble: 1 × 6
#>      id   pid  ppid name        rss       vms
#>   <dbl> <int> <int> <chr> <bch:byt> <bch:byt>
#> 1     1  5820  1612 R         249MB     908MB

# results from syrup are more---or maybe only---useful when
# computations are evaluated in parallel. see package README
# for an example.